Patch GOT at load time
When we compile the rust programs with PIE, the compiler creates the
Global Offset Table (GOT) to put the address of the extern variables.
The GOT is supposed to be fixed at program load time by the dynamic
loader. However, we do not have a dynamic loader and therefore, the GOT
entries are un-patched and contain absolute addresses. This causes
problem when the program is triggered in the kernel -- the use of
absolute address will cause the code going to non-existing pages.

Add a new GOT fix step when the base program is loaded.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>