bpf,iu: Fix refcnt leak on base program
The bpf_prog_load_iu function takes the refcnt of the base program but
does not store the base program pointer. This causes that refcnt to be
permanently lost since the pointer is not found during
bpf_prog_free_deferred.

Store the pointer in the newly created bpf_prog struct so that the
refcnt will be decremented correctly.

Fixes: 73e6a00a2bbe ("implement subprog-loading functions to support multiple programs in the same file")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>