rex: force loader process to dump core upon panic
Send a SIGSYS to the loader of the program if the program panics and
unwinds. This effectively kills the loader if there is no handler
installed for SIGSYS (regarded as intentionally ignoring the panic),
which in turn will remove the program reference from the loader. Though
other references to the program are not removed (e.g., reference from
pinning and prog-array maps), this should still serve as an effective
notification to userspace that something is wrong with the program.
Alternatively, if userspace knows how to recover, it can install a
handler for SIGSYS and perform failure recovery upon the panic.

Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>