x86/rex: fix OOB access on rex stack area
The mapped area for rex stack is [va, va + REX_STACK_SIZE), where va
comes from the call to vmap() and is the start of the virtually mapped
stack. However, we previously set rex_stack_ptr to va + REX_STACK_SIZE,
which is 1 byte outside this range. It is also weird that we have never
seen a guard page hit, probably because this 1 byte is never accessed
somehow.

Use rex_stack_ptr = va + REX_STACK_SIZE - 8 to keep the stack pointer
within the virtually mapped stack and 8-byte-aligned.

Fixes: 8f1b74f37100 ("allocate vmapped per-cpu iu_stack")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>