rex: do not count total memory size and page counts incermentally
We previously counted the total memory and page counts needed for the
program incrementally. This causes problems when the linker (e.g. mold)
generates a gap page between LOAD segments, as that gap page will not be
counted.

Instead, directly calculate the total memory and page counts by aligning
the largest memeory address found in the LOAD segments to page boundary.

Fixes: 88b2c24dc7b0 ("Fixed memory conflict for distributed apps")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>